+++ src/Microsoft.DotNet.XHarness.Apple/Orchestration/RunOrchestrator.cs
@@ -4,34 +4,24 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.DotNet.XHarness.Common.CLI;
-using Microsoft.DotNet.XHarness.Common.Execution;
 using Microsoft.DotNet.XHarness.Common.Logging;
 using Microsoft.DotNet.XHarness.iOS.Shared;
 using Microsoft.DotNet.XHarness.iOS.Shared.Execution;
 using Microsoft.DotNet.XHarness.iOS.Shared.Hardware;
 using Microsoft.DotNet.XHarness.iOS.Shared.Logging;
-using Microsoft.DotNet.XHarness.iOS.Shared.Utilities;
 
 namespace Microsoft.DotNet.XHarness.Apple
 {
     /// <summary>
     /// This orchestrator implements the `run` command flow.
-    /// In this flow we spawn the application and do not expect TestRunner inside.
+    /// In this flow we install, run and uninstall the application and do not expect TestRunner inside.
     /// We only try to detect the exit code after the app run is finished.
     /// </summary>
-    public class RunOrchestrator : BaseOrchestrator
+    public class RunOrchestrator : BaseRunOrchestrator
     {
-        private readonly IMlaunchProcessManager _processManager;
-        private readonly ILogger _logger;
-        private readonly ILogs _logs;
-        private readonly IFileBackedLog _mainLog;
-        private readonly IErrorKnowledgeBase _errorKnowledgeBase;
-        private readonly AppRunner _appRunner;
-
         public RunOrchestrator(
             IMlaunchProcessManager processManager,
             IAppBundleInformationParser appBundleInformationParser,
@@ -39,26 +29,9 @@ namespace Microsoft.DotNet.XHarness.Apple
             ILogger logger,
             ILogs logs,
             IFileBackedLog mainLog,
-            IErrorKnowledgeBase errorKnowledgeBase) : base(processManager, appBundleInformationParser, deviceFinder, logger, mainLog, errorKnowledgeBase)
+            IErrorKnowledgeBase errorKnowledgeBase)
+            : base(processManager, appBundleInformationParser, deviceFinder, logger, logs, mainLog, errorKnowledgeBase)
         {
-            _processManager = processManager ?? throw new ArgumentNullException(nameof(processManager));
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-            _logs = logs ?? throw new ArgumentNullException(nameof(logs));
-            _mainLog = mainLog ?? throw new ArgumentNullException(nameof(mainLog));
-            _errorKnowledgeBase = errorKnowledgeBase ?? throw new ArgumentNullException(nameof(errorKnowledgeBase));
-
-            // Only add the extra callback if we do know that the feature was indeed enabled
-            Action<string>? logCallback = IsLldbEnabled() ? (l) => NotifyUserLldbCommand(_logger, l) : null;
-
-            _appRunner = new AppRunner(
-                _processManager,
-                new CrashSnapshotReporterFactory(_processManager),
-                new CaptureLogFactory(),
-                new DeviceLogCapturerFactory(_processManager),
-                _mainLog,
-                _logs,
-                new Helpers(),
-                logCallback);
         }
 
         public Task<ExitCode> OrchestrateRun(
@@ -103,100 +76,6 @@ namespace Microsoft.DotNet.XHarness.Apple
                 executeMacCatalystApp,
                 executeApp,
                 cancellationToken);
-        }
-
-        private async Task<ExitCode> ExecuteApp(
-            AppBundleInformation appBundleInfo,
-            TestTargetOs target,
-            IDevice device,
-            IDevice? companionDevice,
-            TimeSpan timeout,
-            int expectedExitCode,
-            IReadOnlyCollection<(string, string)> environmentalVariables,
-            IEnumerable<string> passthroughArguments,
-            CancellationToken cancellationToken)
-        {
-            _logger.LogInformation($"Starting application '{appBundleInfo.AppName}' on '{device.Name}'");
-
-            ProcessExecutionResult result = await _appRunner.RunApp(
-                appBundleInfo,
-                target,
-                device,
-                companionDevice,
-                timeout,
-                passthroughArguments,
-                environmentalVariables,
-                cancellationToken);
-
-            if (!result.Succeeded)
-            {
-                _logger.LogError($"App run has failed. mlaunch exited with {result.ExitCode}");
-                return ExitCode.APP_LAUNCH_FAILURE;
-            }
-
-            return ParseResult(new iOSExitCodeDetector(), expectedExitCode, appBundleInfo, result);
-        }
-
-        private async Task<ExitCode> ExecuteMacCatalystApp(
-            AppBundleInformation appBundleInfo,
-            TimeSpan timeout,
-            int expectedExitCode,
-            IReadOnlyCollection<(string, string)> environmentalVariables,
-            IEnumerable<string> passthroughArguments,
-            CancellationToken cancellationToken)
-        {
-            _logger.LogInformation($"Starting '{appBundleInfo.AppName}' on MacCatalyst");
-
-            ProcessExecutionResult result = await _appRunner.RunMacCatalystApp(
-                appBundleInfo,
-                timeout,
-                passthroughArguments,
-                environmentalVariables,
-                cancellationToken: cancellationToken);
-
-            return ParseResult(new MacCatalystExitCodeDetector(), expectedExitCode, appBundleInfo, result);
-        }
-
-        private ExitCode ParseResult(
-            IExitCodeDetector exitCodeDetector,
-            int expectedExitCode,
-            AppBundleInformation appBundleInfo,
-            ProcessExecutionResult result)
-        {
-            if (result.TimedOut)
-            {
-                _logger.LogError($"App run has timed out");
-                return ExitCode.TIMED_OUT;
-            }
-
-            int exitCode;
-
-            var systemLog = _logs.FirstOrDefault(log => log.Description == LogType.SystemLog.ToString());
-            if (systemLog == null)
-            {
-                _logger.LogError("Application has finished but no system log found. Failed to determine the exit code!");
-                return ExitCode.RETURN_CODE_NOT_SET;
-            }
-
-            exitCode = exitCodeDetector.DetectExitCode(appBundleInfo, systemLog);
-            _logger.LogInformation($"App run ended with {exitCode}");
-
-            if (expectedExitCode != exitCode)
-            {
-                _logger.LogError($"Application has finished with exit code {exitCode} but {expectedExitCode} was expected");
-
-                if (_errorKnowledgeBase.IsKnownTestIssue(_mainLog, out var failureMessage))
-                {
-                    _logger.LogError(failureMessage.Value.HumanMessage);
-                }
-
-                return ExitCode.GENERAL_FAILURE;
-            }
-
-            _logger.LogInformation("Application has finished with exit code: " + exitCode +
-                (expectedExitCode != 0 ? " (as expected)" : null));
-
-            return ExitCode.SUCCESS;
         }
     }
 }
